NDLA-417: Refactored elasticsearch indexer
