+++ src/main/scala/no/ndla/articleapi/service/search/ElasticContentIndexComponent.scala
@@ -7,7 +7,7 @@
  */
 
 
-package no.ndla.articleapi.service
+package no.ndla.articleapi.service.search
 
 import java.text.SimpleDateFormat
 import java.util.Calendar
